maj globale
fichiers corrects

+++ admin/delete_user.php
@@ -1,13 +1,15 @@
 <?php
+
 require_once 'inc/session.php';
 require_once 'inc/connect.php';
-if(!$is_logged){
-    header('Location: index.php');
-}
 
 $user = null;
 if(isset($_GET['id']) && is_numeric($_GET['id'])){
-
+    if($_SESSION['perm'] < 2 || $_SESSION['id'] != $_GET['id']){
+        header('Location: admin_main.php');
+        die();
+    }
+    
 	if(!empty($_POST)) {
 		if(isset($_POST['delete'])){
 			$delete = $bdd->prepare('DELETE FROM users WHERE id = :idUser');
@@ -35,31 +37,32 @@ if(isset($_GET['id']) && is_numeric($_GET['id'])){
 <head>
 	<meta charset="utf-8">
 	<title>Supprimer un utilisateur</title>
+	<link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+    <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
+    <link rel="stylesheet" href="style/main.css">
 </head>
 <body>
 	
 	<?php include_once 'inc/navbar.php' ?>
-	<main class="container">
-    <?php if($_SESSION['perm'] >=2 || $_SESSION['id'] === $_GET['id']): ?>
-        <h1>Supprimer un utilisateur</h1>
+	
+	<h1>Supprimer un utilisateur</h1>
+	
+	<hr>
 
-        <hr>
+	<?php if(empty($user)): ?>
+		<div class="alert alert-danger">
+			Utilisateur inconnu !
+		</div>
+	<?php else: ?>
+		<p>Voulez-vous vraiment supprimer l'utilisateur &laquo; <?=$user['firstname'];?> &raquo; ?</p>
 
-        <?php if(empty($user)): ?>
-            <div class="alert alert-danger">
-                Utilisateur inconnu !
-            </div>
-        <?php else: ?>
-            <p>Voulez-vous vraiment supprimer l'utilisateur &laquo; <?=$user['firstname'];?> &raquo; ?</p>
-
-            <form method="POST">
-                <!-- Annuler -->
-                <input type="button" onclick="history.back();" value="Annuler" class="btn btn-default">
-                <!-- Valider -->
-                <input type="submit" name="delete" value="Oui, je veux supprimer cet utilisateur">
-            </form>
-        <?php endif;?>
-    <?php endif; ?>
-	</main>
+		<form method="POST">
+			<!-- Annuler -->
+			<input type="button" onclick="history.back();" value="Annuler" class="btn btn-default">
+			<!-- Valider -->
+			<input type="submit" name="delete" value="Oui, je veux supprimer cet utilisateur">
+		</form>
+	<?php endif;?>
+	
 </body>
 </html>
